refactor : BoardMaker에서 하던 역할을 Board로 이동
