+++ src/main/kotlin/domain/BoardMaker.kt
@@ -6,10 +6,8 @@ package domain
  */
 class BoardMaker(private val gameSettingInfo: GameSettingInfo) {
     fun makeBoard(manualMines: List<Position>? = null): Board {
-        val board = Board(createGround())
-        board.setAutoMine(gameSettingInfo.mineCount)
+        val board = Board(createGround(), gameSettingInfo)
         board.setManualMine(manualMines)
-        setMineCount(board)
         return board
     }
 
@@ -19,31 +17,10 @@ class BoardMaker(private val gameSettingInfo: GameSettingInfo) {
                 Position(width, height)
             }
         }
-
         return positions.associateWith { Ground() }
-    }
-
-    private fun setMineCount(board: Board) {
-        val minePositions = board.grounds.filter { it.value.isMine }.map { it.key }
-        minePositions.forEach {
-            setMineCount(it, board)
-        }
-    }
-
-    private fun setMineCount(position: Position, board: Board) {
-        repeat(ARROW_COUNT) {
-            val tempX = position.x + X_MARGINS[it]
-            val tempY = position.y + Y_MARGINS[it]
-            val position = Position.makePositionOrNull(tempX, tempY, gameSettingInfo.width, gameSettingInfo.height)
-                ?: return@repeat
-            board.grounds[position]?.addMineCount()
-        }
     }
 
     companion object {
         private const val START_INDEX = 0
-        const val ARROW_COUNT = 8
-        val Y_MARGINS = listOf(-1, -1, -1, 0, 0, 1, 1, 1)
-        val X_MARGINS = listOf(-1, 0, 1, -1, 1, -1, 0, 1)
     }
 }
